+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This release is a modernization effort. It adds support for Alliance Auth V5 and
 
 - HomeStationFilter not working correctly
 - AltFilter not returning audit data for non-matching users
+- ActivityFilter not returning audit data for non-matching users
 
 ## [1.6.1] - 2025-11-04